Move common package out of gateway package.
