+++ common/pointers.go
