+++ gateways/authorize_net/authorizenet.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/BoltApp/sleet"
-	"github.com/BoltApp/sleet/gateways/common"
+	"github.com/BoltApp/sleet/common"
 	"io/ioutil"
 	"net/http"
 )
